@@ -3,7 +3,7 @@ Get all email addresses in a string.
 
 @example
 ```
-import getEmails = require('get-emails');
+import getEmails from 'get-emails';
 
 const text = 'Lorem ipsum dolor, sindresorhus@gmail.com consectetuer unicorn@rainbow.cake elit.';
 
@@ -11,6 +11,4 @@ getEmails(text);
 //=> Set {'sindresorhus@gmail.com', 'unicorn@rainbow.cake'}
 ```
 */
-declare function getEmails(string: string): Set<string>;
-
-export = getEmails;
+export default function getEmails(string: string): Set<string>;
