@@ -1,8 +1,8 @@
 import {expectType} from 'tsd';
-import getEmails = require('.');
+import getEmails from './index.js';
 
 expectType<Set<string>>(
 	getEmails(
-		'Lorem ipsum dolor, sindresorhus@gmail.com consectetuer unicorn@rainbow.cake elit.'
-	)
+		'Lorem ipsum dolor, sindresorhus@gmail.com consectetuer unicorn@rainbow.cake elit.',
+	),
 );
